climbingTicks: refactor - move to ticks folder
