+++ src/components/FeaturePanel/Climbing/Ticks/TickStyleSelect.tsx
@@ -3,7 +3,7 @@ import { FormControl, MenuItem, Select, Tooltip } from '@mui/material';
 import styled from '@emotion/styled';
 import InfoOutlinedIcon from '@mui/icons-material/InfoOutlined';
 import { TickStyle } from '../types';
-import { tickStyles } from '../../../../services/ticks';
+import { tickStyles } from '../../../../services/my-ticks/ticks';
 
 const Row = styled.div`
   display: flex;
